+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.1] - 2019-05-08
+### Changed
+- Initializer of the alert changed to not receive corner radius.
+
+### Fixed
+ - Number of lines of the label.
+
 ## [1.0.0] - 2019-04-17
 ### Added
 - First visual componnents of alert.
@@ -13,4 +20,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support to bottom position of alert.
 
 [Unreleased]: https://github.com/cardoso19/MDTAlert
+[1.0.1]: https://github.com/cardoso19/MDTAlert/releases/tag/v1.0.1
 [1.0.0]: https://github.com/cardoso19/MDTAlert/releases/tag/v1.0.0